// apps/storeapp/src/api-client/store/models.ts
export type StoreRoleData = {
  payloads: {
    List: {
      authorization?: string;

      query: {
        page: string;
        size: string;
      };
    };
    Create: {
      authorization?: string;
      requestBody: {
        name: string;
        permissions: unknown;
        super_admin?: boolean;
        order: string | number;
        color: string;
      };

      query?: {};
    };
    Update: {
      authorization?: string;
      requestBody: {
        name: string;
        permissions: unknown;
        super_admin?: boolean;
        order: string | number;
        color: string;
      };

      query: {
        id: string;
      };
    };
    Delete: {
      authorization?: string;

      query: {
        id: string;
      };
    };
  };

  responses: {
    List: {
      status: boolean;
      message: string;
      data: Array<{
        _id: string;
        name: string;
        permissions: Record<string, unknown>;
        order: string | number;
        super_admin: boolean;
        color: string;
      }>;
      meta: {
        pages: number;
        total: number;
        page: number;
        size: number;
      };
    };
    Create: {
      status: boolean;
      message: string;
      data: {
        _id: string;
        name: string;
        permissions: Record<string, unknown>;
        order: string | number;
        super_admin: boolean;
        color: string;
      };
    };
    Update: {
      status: boolean;
      message: string;
      data: {
        _id: string;
        name: string;
        permissions: Record<string, unknown>;
        order: string | number;
        super_admin: boolean;
        color: string;
      };
    };
    Delete: {
      status: boolean;
      message: string;
      data: {
        _id: string;
        name: string;
        permissions: Record<string, unknown>;
        order: string | number;
        super_admin: boolean;
        color: string;
      };
    };
  };
};